Conflicts:

	release/src/router/dnsmasq/CHANGELOG
	release/src/router/dnsmasq/Makefile
	release/src/router/dnsmasq/bld/Android.mk
	release/src/router/dnsmasq/man/dnsmasq.8
	release/src/router/dnsmasq/man/fr/dnsmasq.8
	release/src/router/dnsmasq/po/de.po
	release/src/router/dnsmasq/po/es.po
	release/src/router/dnsmasq/po/fi.po
	release/src/router/dnsmasq/po/fr.po
	release/src/router/dnsmasq/po/id.po
	release/src/router/dnsmasq/po/it.po
	release/src/router/dnsmasq/po/no.po
	release/src/router/dnsmasq/po/pl.po
	release/src/router/dnsmasq/po/pt_BR.po
	release/src/router/dnsmasq/po/ro.po
	release/src/router/dnsmasq/src/config.h
	release/src/router/dnsmasq/src/conntrack.c
	release/src/router/dnsmasq/src/dnsmasq.c
	release/src/router/dnsmasq/src/dnsmasq.h
	release/src/router/dnsmasq/src/lease.c
	release/src/router/dnsmasq/src/netlink.c
	release/src/router/dnsmasq/src/option.c
	release/src/router/dnsmasq/src/rfc2131.c

+++ release/src/router/dnsmasq/src/rfc1035.c
@@ -1,4 +1,4 @@
-/* dnsmasq is Copyright (c) 2000-2011 Simon Kelley
+/* dnsmasq is Copyright (c) 2000-2012 Simon Kelley
 
    This program is free software; you can redistribute it and/or modify
    it under the terms of the GNU General Public License as published by
@@ -25,7 +25,7 @@ static int add_resource_record(struct dns_header *header, char *limit, int *trun
     ((size_t)((pp) - (unsigned char *)(header) + (len)) <= (plen))
 
 #define ADD_RDLEN(header, pp, plen, len) \
-    (!CHECK_LEN(header, pp, plen, len) ? 0 : (long)((pp) += (len)), 1)
+  (!CHECK_LEN(header, pp, plen, len) ? 0 : (((pp) += (len)), 1))
 
 static int extract_name(struct dns_header *header, size_t plen, unsigned char **pp, 
 			char *name, int isExtract, int extrabytes)
